Refactor project structure to move files to appropriate directories
Move files from the root directory to appropriate subdirectories to maintain a clean project structure.

* Move `appsettings.json` to `config` directory.
* Move `SynthesisRequest.cs`, `SynthesisResult.cs`, `ResearchResult.cs`, `ResearchSource.cs`, and `KnowledgeWorkController.cs` to `src/BusinessApplications/ResearchAnalysis` directory.
* Move `SelfEvaluator.cs` and `MetacognitiveOversightComponent.cs` to `src/MetacognitiveLayer/SelfEvaluation` directory.
* Move `SemanticSearchManager.cs` and `EnhancedRAGSystem.cs` to `src/FoundationLayer/SemanticSearch` directory.
* Move `SituationAnalysisRequest.cs`, `SituationAnalysisResponse.cs`, `OptionsGenerationRequest.cs`, and `DecisionSupportController.cs` to `src/BusinessApplications/DecisionSupport` directory.
* Move `TroubleshootingRequest.cs`, `TroubleshootingResponse.cs`, `UserFeedback.cs`, `InteractionRecord.cs`, `ProductInfo.cs`, and `CustomerServiceController.cs` to `src/BusinessApplications/CustomerIntelligence` directory.
* Move `UncertaintyQuantifier.cs` to `src/MetacognitiveLayer/UncertaintyQuantification` directory.
* Move `VectorDatabaseManager.cs` to `src/FoundationLayer/VectorDatabase` directory.
* Move `KnowledgeManager.cs` and `KnowledgeDocument.cs` to `src/BusinessApplications/KnowledgeManagement` directory.
* Move `LearningManager.cs`, `LearningInsight.cs`, and `ContinuousLearningComponent.cs` to `src/MetacognitiveLayer/ContinuousLearning` directory.
* Move `MultiPerspectiveCognition.cs` to `src/ReasoningLayer/AnalyticalReasoning` directory.
* Move `EthicalReasoner.cs` to `src/ReasoningLayer/EthicalReasoning` directory.
* Move `FeatureFlagManager.cs` to `src/FoundationLayer/EnterpriseConnectors` directory.
* Move `DocumentIngestionFunction.cs` to `src/FoundationLayer/DocumentProcessing` directory.
* Move `DomainSpecificReasoner.cs` to `src/ReasoningLayer/DomainSpecificReasoning` directory.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/phoenixvc/cognitive-mesh?shareId=XXXX-XXXX-XXXX-XXXX).
