--- ResearchSource.cs
@@ -1,6 +1,0 @@
-public class ResearchSource
-{
-    public string Title { get; set; }
-    public string Source { get; set; }
-    public string Snippet { get; set; }
-}
