--- SemanticSearchManager.cs
@@ -1,205 +1,0 @@
-using Azure.AI.OpenAI;
-using Azure.Search.Documents;
-using Azure.Search.Documents.Models;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-public class SemanticSearchManager
-{
-    private readonly SearchClient _searchClient;
-    private readonly OpenAIClient _openAIClient;
-    private readonly string _embeddingDeployment;
-    private readonly string _completionDeployment;
-    private readonly FeatureFlagManager _featureFlagManager;
-
-    public SemanticSearchManager(SearchClient searchClient, OpenAIClient openAIClient, string embeddingDeployment, string completionDeployment, FeatureFlagManager featureFlagManager)
-    {
-        _searchClient = searchClient;
-        _openAIClient = openAIClient;
-        _embeddingDeployment = embeddingDeployment;
-        _completionDeployment = completionDeployment;
-        _featureFlagManager = featureFlagManager;
-    }
-
-    public async Task<List<KnowledgeDocument>> PerformSemanticSearchAsync(string query, int limit = 5, string filter = null)
-    {
-        if (_featureFlagManager.EnableADK)
-        {
-            // Implement logic for ADK framework
-            return await PerformADKSemanticSearchAsync(query, limit, filter);
-        }
-        else if (_featureFlagManager.EnableLangGraph)
-        {
-            // Implement logic for LangGraph framework
-            return await PerformLangGraphSemanticSearchAsync(query, limit, filter);
-        }
-        else if (_featureFlagManager.EnableCrewAI)
-        {
-            // Implement logic for CrewAI framework
-            return await PerformCrewAISemanticSearchAsync(query, limit, filter);
-        }
-        else if (_featureFlagManager.EnableSemanticKernel)
-        {
-            // Implement logic for Semantic Kernel framework
-            return await PerformSemanticKernelSemanticSearchAsync(query, limit, filter);
-        }
-        else if (_featureFlagManager.EnableAutoGen)
-        {
-            // Implement logic for AutoGen framework
-            return await PerformAutoGenSemanticSearchAsync(query, limit, filter);
-        }
-        else if (_featureFlagManager.EnableSmolagents)
-        {
-            // Implement logic for Smolagents framework
-            return await PerformSmolagentsSemanticSearchAsync(query, limit, filter);
-        }
-        else if (_featureFlagManager.EnableAutoGPT)
-        {
-            // Implement logic for AutoGPT framework
-            return await PerformAutoGPTSemanticSearchAsync(query, limit, filter);
-        }
-        else
-        {
-            return new List<KnowledgeDocument> { new KnowledgeDocument { Title = "Feature not enabled.", Content = "No framework is enabled for semantic search." } };
-        }
-    }
-
-    private async Task<List<KnowledgeDocument>> PerformADKSemanticSearchAsync(string query, int limit, string filter)
-    {
-        // Implement ADK-specific semantic search logic
-        return await Task.FromResult(new List<KnowledgeDocument>());
-    }
-
-    private async Task<List<KnowledgeDocument>> PerformLangGraphSemanticSearchAsync(string query, int limit, string filter)
-    {
-        // Implement LangGraph-specific semantic search logic
-        return await Task.FromResult(new List<KnowledgeDocument>());
-    }
-
-    private async Task<List<KnowledgeDocument>> PerformCrewAISemanticSearchAsync(string query, int limit, string filter)
-    {
-        // Implement CrewAI-specific semantic search logic
-        return await Task.FromResult(new List<KnowledgeDocument>());
-    }
-
-    private async Task<List<KnowledgeDocument>> PerformSemanticKernelSemanticSearchAsync(string query, int limit, string filter)
-    {
-        // Generate embedding for query
-        var embeddingResponse = await _openAIClient.GetEmbeddingsAsync(
-            _embeddingDeployment,
-            new EmbeddingsOptions(query));
-
-        float[] queryEmbedding = embeddingResponse.Value.Data[0].Embedding.ToArray();
-
-        // Create vector query
-        var vectorQuery = new VectorizedQuery(queryEmbedding)
-        {
-            KNearestNeighborsCount = limit,
-            Fields = { "embedding" }
-        };
-
-        // Execute search
-        var searchOptions = new SearchOptions
-        {
-            VectorSearch = new VectorSearchOptions
-            {
-                Queries = { vectorQuery }
-            },
-            Size = limit
-        };
-
-        // Add filter if provided
-        if (!string.IsNullOrEmpty(filter))
-        {
-            searchOptions.Filter = filter;
-        }
-
-        var searchResults = await _searchClient.SearchAsync<SearchDocument>(null, searchOptions);
-
-        // Process results
-        var documents = new List<KnowledgeDocument>();
-        await foreach (var result in searchResults.GetResultsAsync())
-        {
-            var doc = new KnowledgeDocument
-            {
-                Id = result.Document["id"].ToString(),
-                Title = result.Document["title"].ToString(),
-                Content = result.Document["content"].ToString(),
-                Source = result.Document["source"].ToString(),
-                Category = result.Document["category"].ToString(),
-                Created = (DateTimeOffset)result.Document["created"],
-                Tags = ((string[])result.Document["tags"]).ToList()
-            };
-
-            documents.Add(doc);
-        }
-
-        return documents;
-    }
-
-    private async Task<List<KnowledgeDocument>> PerformAutoGenSemanticSearchAsync(string query, int limit, string filter)
-    {
-        // Implement AutoGen-specific semantic search logic
-        return await Task.FromResult(new List<KnowledgeDocument>());
-    }
-
-    private async Task<List<KnowledgeDocument>> PerformSmolagentsSemanticSearchAsync(string query, int limit, string filter)
-    {
-        // Implement Smolagents-specific semantic search logic
-        return await Task.FromResult(new List<KnowledgeDocument>());
-    }
-
-    private async Task<List<KnowledgeDocument>> PerformAutoGPTSemanticSearchAsync(string query, int limit, string filter)
-    {
-        // Implement AutoGPT-specific semantic search logic
-        return await Task.FromResult(new List<KnowledgeDocument>());
-    }
-
-    public async Task<string> GenerateResponseWithSemanticSearchAsync(string query, string systemPrompt = null)
-    {
-        // Perform semantic search
-        var documents = await PerformSemanticSearchAsync(query, 5);
-
-        if (documents.Count == 0)
-        {
-            return "I don't have enough information to answer that question.";
-        }
-
-        // Format documents as context
-        var context = new StringBuilder();
-        foreach (var doc in documents)
-        {
-            context.AppendLine($"--- Document: {doc.Title} ---");
-            context.AppendLine($"Source: {doc.Source}");
-            context.AppendLine(doc.Content);
-            context.AppendLine();
-        }
-
-        // Create system prompt if not provided
-        if (string.IsNullOrEmpty(systemPrompt))
-        {
-            systemPrompt = "You are a helpful assistant that answers questions based on the provided context. " +
-                          "If the answer cannot be found in the context, say that you don't know.";
-        }
-
-        // Create completion options
-        var chatCompletionOptions = new ChatCompletionsOptions
-        {
-            DeploymentName = _completionDeployment,
-            Temperature = 0.3f,
-            MaxTokens = 800,
-            Messages =
-            {
-                new ChatRequestSystemMessage(systemPrompt),
-                new ChatRequestUserMessage($"Context:\n{context}\n\nQuestion: {query}")
-            }
-        };
-
-        // Generate response
-        var response = await _openAIClient.GetChatCompletionsAsync(chatCompletionOptions);
-
-        return response.Value.Choices[0].Message.Content;
-    }
-}
