--- SituationAnalysisRequest.cs
@@ -1,5 +1,0 @@
-public class SituationAnalysisRequest
-{
-    public string Situation { get; set; }
-    public List<string> Perspectives { get; set; }
-}
