--- SynthesisResult.cs
@@ -1,6 +1,0 @@
-public class SynthesisResult
-{
-    public string FocusArea { get; set; }
-    public string Synthesis { get; set; }
-    public string KeyTakeaways { get; set; }
-}
