--- TroubleshootingRequest.cs
@@ -1,6 +1,0 @@
-public class TroubleshootingRequest
-{
-    public string ProductId { get; set; }
-    public string IssueDescription { get; set; }
-    public string PreviousStepsTaken { get; set; }
-}
