--- UserFeedback.cs
@@ -1,6 +1,0 @@
-public class UserFeedback
-{
-    public string QueryId { get; set; }
-    public int Rating { get; set; }
-    public string Comments { get; set; }
-}
