--- VectorDatabaseManager.cs
@@ -1,90 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Threading.Tasks;
-using Microsoft.Extensions.Logging;
-
-public class VectorDatabaseManager
-{
-    private readonly ILogger<VectorDatabaseManager> _logger;
-    private readonly FeatureFlagManager _featureFlagManager;
-
-    public VectorDatabaseManager(ILogger<VectorDatabaseManager> logger, FeatureFlagManager featureFlagManager)
-    {
-        _logger = logger;
-        _featureFlagManager = featureFlagManager;
-    }
-
-    public async Task<bool> StoreVectorAsync(string documentId, float[] vector)
-    {
-        try
-        {
-            if (_featureFlagManager.EnableADK)
-            {
-                _logger.LogInformation($"Storing vector for document: {documentId}");
-
-                // Simulate vector storage logic
-                await Task.Delay(1000);
-
-                _logger.LogInformation($"Successfully stored vector for document: {documentId}");
-                return true;
-            }
-            else
-            {
-                _logger.LogInformation("Feature not enabled.");
-                return false;
-            }
-        }
-        catch (Exception ex)
-        {
-            _logger.LogError($"Failed to store vector for document: {documentId}. Error: {ex.Message}");
-            return false;
-        }
-    }
-
-    public async Task<float[]> RetrieveVectorAsync(string documentId)
-    {
-        try
-        {
-            if (_featureFlagManager.EnableLangGraph)
-            {
-                _logger.LogInformation($"Retrieving vector for document: {documentId}");
-
-                // Simulate vector retrieval logic
-                await Task.Delay(1000);
-
-                var vector = new float[] { 0.1f, 0.2f, 0.3f, 0.4f, 0.5f };
-
-                _logger.LogInformation($"Successfully retrieved vector for document: {documentId}");
-                return vector;
-            }
-            else
-            {
-                _logger.LogInformation("Feature not enabled.");
-                return null;
-            }
-        }
-        catch (Exception ex)
-        {
-            _logger.LogError($"Failed to retrieve vector for document: {documentId}. Error: {ex.Message}");
-            throw;
-        }
-    }
-
-    public async Task IntegrateWithFabricDataMeshAsync(string documentId, float[] vector)
-    {
-        try
-        {
-            _logger.LogInformation($"Integrating vector for document: {documentId} with Fabric data mesh");
-
-            // Simulate integration with Fabric data mesh
-            await Task.Delay(1000);
-
-            _logger.LogInformation($"Successfully integrated vector for document: {documentId} with Fabric data mesh");
-        }
-        catch (Exception ex)
-        {
-            _logger.LogError($"Failed to integrate vector for document: {documentId} with Fabric data mesh. Error: {ex.Message}");
-            throw;
-        }
-    }
-}
