--- appsettings.json
@@ -1,44 +1,0 @@
-{
-  "FeatureFlags": {
-    "UseOneLake": true,
-    "enable_multi_agent": true,
-    "enable_dynamic_task_routing": true,
-    "enable_stateful_workflows": true,
-    "enable_human_in_the_loop": false,
-    "enable_tool_integration": true,
-    "enable_memory_management": true,
-    "enable_streaming": false,
-    "enable_code_execution": true,
-    "enable_guardrails": true,
-    "enable_enterprise_integration": true,
-    "enable_modular_skills": true,
-    "enable_ADK": true,
-    "enable_ADK_WorkflowAgents": true,
-    "enable_ADK_ToolIntegration": false,
-    "enable_ADK_Guardrails": true,
-    "enable_ADK_Multimodal": false,
-    "enable_LangGraph": true,
-    "enable_LangGraph_Stateful": true,
-    "enable_LangGraph_Streaming": false,
-    "enable_LangGraph_HITL": true,
-    "enable_CrewAI": true,
-    "enable_CrewAI_Team": true,
-    "enable_CrewAI_DynamicPlanning": true,
-    "enable_CrewAI_AdaptiveExecution": true,
-    "enable_SemanticKernel": true,
-    "enable_SemanticKernel_Memory": true,
-    "enable_SemanticKernel_Security": true,
-    "enable_SemanticKernel_Automation": true,
-    "enable_AutoGen": true,
-    "enable_AutoGen_Conversations": true,
-    "enable_AutoGen_Context": true,
-    "enable_AutoGen_APIIntegration": true,
-    "enable_Smolagents": true,
-    "enable_Smolagents_Modular": true,
-    "enable_Smolagents_Context": true,
-    "enable_AutoGPT": true,
-    "enable_AutoGPT_Autonomous": true,
-    "enable_AutoGPT_Memory": true,
-    "enable_AutoGPT_InternetAccess": true
-  }
-}
