+++ src/BusinessApplications/CustomerIntelligence/InteractionRecord.cs
