+++ src/BusinessApplications/CustomerIntelligence/ProductInfo.cs
