+++ src/BusinessApplications/KnowledgeManagement/KnowledgeDocument.cs
