+++ src/BusinessApplications/KnowledgeManagement/KnowledgeManager.cs
