+++ src/FoundationLayer/DocumentProcessing/DocumentIngestionFunction.cs
