+++ src/FoundationLayer/EnterpriseConnectors/FeatureFlagManager.cs
