+++ src/MetacognitiveLayer/ContinuousLearning/ContinuousLearningComponent.cs
