+++ src/MetacognitiveLayer/ContinuousLearning/LearningInsight.cs
