+++ src/MetacognitiveLayer/ContinuousLearning/LearningManager.cs
