+++ src/ReasoningLayer/AnalyticalReasoning/MultiPerspectiveCognition.cs
